branch microservices Some fixes Extracting a main-service and request-service as a microservice Enable Feign clients All local Postman tests passed successfully

+++ core/category-service/src/main/java/ru/yandex/practicum/service/CategoryServiceImpl.java
@@ -1,5 +1,6 @@
 package ru.yandex.practicum.service;
 
+import feign.FeignException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -7,6 +8,8 @@ import org.springframework.stereotype.Service;
 import ru.yandex.practicum.dto.NewCategoryDto;
 import ru.yandex.practicum.dto.category.CategoryDto;
 import ru.yandex.practicum.exception.EntityNotFoundException;
+import ru.yandex.practicum.exception.InitiatorRequestException;
+import ru.yandex.practicum.feignClient.EventClient;
 import ru.yandex.practicum.mapper.CategoryMapper;
 import ru.yandex.practicum.model.Category;
 import ru.yandex.practicum.repository.CategoryRepository;
@@ -18,6 +21,7 @@ import java.util.List;
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
+    private final EventClient eventClient;
 
     @Override
     public List<CategoryDto> getAll(Integer from, Integer size) {
@@ -42,6 +46,11 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteCategory(Long id) {
+        Category category = categoryRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException(Category.class, "Категория с ID - " + id + ", не найдена."));
+        if (findEventsWithCategory(id)) {
+            throw new InitiatorRequestException("По Категории с ID: " + category.getId() + " найдены события.");
+        }
         categoryRepository.deleteById(id);
     }
 
@@ -57,4 +66,13 @@ public class CategoryServiceImpl implements CategoryService {
     public List<CategoryDto> getCategorysList(List<Long> ids) {
         return categoryMapper.toCategoryDtoList(categoryRepository.findAllById(ids));
     }
+
+    private Boolean findEventsWithCategory(Long id) {
+        try {
+            return eventClient.findEventsWithCategory(id);
+        } catch (FeignException e) {
+            throw new EntityNotFoundException(FeignException.class, "События с категорией ID - " + id + ", не найдены.");
+        }
+    }
+
 }
